requirements.txt
gradio
google.genai
dotenv
